start QuickFilters rework
For script quick filters, it seems that scripts and UI can't be saved in
the EntityStore like it was done before. Both Java and Script quick
filters will generate an eclipse plugin which will be installed in the
policy studio.

Java QuickFilters will be generated from annotation processing. Script
QuickFilters will still use an external tool.

Generation will take the form of a jar for the policy studio. The
typeset will need to be extracted from the produced archive to be
imported in the policy studio UI.

Changes:
- move QuickFilters annotations and builder to a dedicated project
- switch to java-library gradle plugin
- add a dummy annotation processor

TODO:
- start to see how it works with yaml.