Refactoring script extensions
Script extensions can now optionally interact with script's resources
and define resources as well. a New method has been added to base bound
runtime to be able to dynamically add extensions to the attaching
script.

Script extensions can either be implemented by declaring and
implementing an interface in a concrete annotated class. To interact
with script resources, an abstract class can be inherited.